requirements.txt
langchain
langgraph
langchain-google-genai
google-generativeai
langchain-chroma
chromadb
pypdf
python-dotenv
pydantic